Set the icon theme from KConfig
If the icon theme is not set, it defaults to hicolor, resulting in icons
rendering as the fallback colored circle and letter.

Paper is used as the default theme name, as this is the most Material-
inspired icon theme available currently for Linux.

- Rename DesktopScrobbler to DesktopFiles and make it a singleton
- Add a property to DesktopFiles to set the icon theme
- Set the icon theme name from the shell settings stored in KConfig
- Make sure the source size of the app icon is always larger than zero
so we don't get loading errors in the console